split test_io.py by function into
- test_get_df_summary.py
- test_get_gnome_pmg_structures.py
- test_load_pymatgen_phonon_docs.py
- test_update_key_name.py
